+++ packages/client-wallets/RELEASES.md
@@ -89,3 +89,13 @@ Based on:
 - [typescript v1.0.0-alpha.10] packages/client-wallets
 ### Releases
 - [NPM v1.0.0-alpha.10] https://www.npmjs.com/package/@formance/sdk-wallets/v/1.0.0-alpha.10 - packages/client-wallets
+
+## 2024-08-14 00:11:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.366.0 (2.396.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.11] packages/client-wallets
+### Releases
+- [NPM v1.0.0-alpha.11] https://www.npmjs.com/package/@formance/sdk-wallets/v/1.0.0-alpha.11 - packages/client-wallets